+++ lib/constants.dart
@@ -25,6 +25,7 @@ const kMaxShownPresets = 10;
 const kMaxNSIPresets = 3;
 const kFollowLinks = false;
 const kUploadOnClose = false;
+const kShowContactSetting = false;
 const kMicromappingTapZoom = 19.0;
 const kManualOption = '✍️';